+++ src/cli/Signal.php
@@ -18,6 +18,8 @@ class Signal
 
     /**
      * Checks exit signals
+     * Used mainly by [[yii\queue\Queue]] to check, whether job execution
+     * loop can be continued.
      * @return bool
      */
     public static function isExit()
@@ -28,7 +30,7 @@ class Signal
             if (!$handled) {
                 foreach ([SIGTERM, SIGINT, SIGHUP] as $signal) {
                     pcntl_signal($signal, function () {
-                        static::$exit = true;
+                        static::setExitFlag();
                     });
                 }
                 $handled = true;
@@ -41,5 +43,15 @@ class Signal
         }
 
         return static::$exit;
+    }
+
+    /**
+     * Sets exit flag to `true`
+     * Method can be used to simulate exit signal for methods that use
+     * [[isExit()]] to check whether execution loop can be continued.
+     */
+    public static function setExitFlag()
+    {
+        static::$exit = true;
     }
 }